Refactored code BuildingNAnimalMenuPatches.cs
